progress on settings
